refactor!: separate understand_ai format into its own directory
