+++ raillabel/format/understand_ai/_translation.py
@@ -82,7 +82,7 @@ def _load_translation():
     global TRANSLATION
 
     translatiion_path = (
-        Path(__file__).parent.parent / "load_" / "loader_classes" / "translation.json"
+        Path(__file__).parent.parent.parent / "load_" / "loader_classes" / "translation.json"
     )
     with translatiion_path.open() as translation_file:
         TRANSLATION = json.load(translation_file)
