+++ raillabel/format/understand_ai/bounding_box_2d.py
@@ -24,7 +24,7 @@ class BoundingBox2d(_Annotation):
     attributes: dict[str, str or list]
         Key value pairs of attributes with the keys beeing the friendly identifier of the
         attribute and the value beeing the attribute value.
-    sensor: raillabel._understand_ai_t4_format.SensorReference
+    sensor: raillabel.format.understand_ai.SensorReference
         Information about the sensor this annotation is labeled in.
     x_min: float
         Left corner of the bounding box in pixels.
