+++ tests/test_raillabel/format/understand_ai/__init__.py
@@ -1,0 +1,2 @@
+# Copyright DB Netz AG and contributors
+# SPDX-License-Identifier: Apache-2.0
