+++ tests/test_raillabel/format/understand_ai/test_uai_frame.py
@@ -10,7 +10,7 @@ import pytest
 
 sys.path.insert(1, str(Path(__file__).parent.parent.parent.parent))
 
-import raillabel._understand_ai_t4_format as uai_format
+import raillabel.format.understand_ai as uai_format
 
 
 def _prepare_frame_data(frame_data: dict, json_data: dict) -> dict:
